Remove an empty file and remove the gov_wifi_authenticatable concern
The code in the concern is only used once and can therefore be merged
with the application_controller.
