+++ app/controllers/application_controller.rb
@@ -3,14 +3,32 @@ class ApplicationController < ActionController::Base
   default_form_builder GOVUKDesignSystemFormBuilder::FormBuilder
 
   include Pagy::Backend
-  include GovWifiAuthenticatable
 
   before_action :redirect_user_with_no_organisation, unless: :devise_controller?
   before_action :update_active_sidebar_path
+  before_action :authenticate_user!, except: :error
+  before_action :confirm_two_factor_setup
+  before_action :configure_devise_permitted_parameters, if: :devise_controller?
   helper_method :current_organisation, :super_admin?
   helper_method :sidebar
   helper_method :subnav
   helper_method :show_navigation_bars
+
+  def error
+    render :error, code: params[:code]
+  end
+
+  def confirm_two_factor_setup
+    return unless current_user &&
+      current_user.need_two_factor_authentication?(request) &&
+      !current_user.totp_enabled?
+
+    redirect_to users_two_factor_authentication_setup_path
+  end
+
+  def configure_devise_permitted_parameters
+    devise_parameter_sanitizer.permit(:accept_invitation, keys: [:name])
+  end
 
   def current_organisation
     if session[:organisation_id] && (current_user.member_of?(session[:organisation_id].to_i) || super_admin?)
